updated changes

+++ sum.py
@@ -3,6 +3,7 @@
 x = input("Type number 1: ")
 y = input("Type number 2: ")
 
-sum = int(x) + int(y)
+# sum = int(x) + int(y)
+sum = float(x) + float(y)
 
 print("The sum is: ", sum)
